Cleaned up client folder, removed unused components and imports, renamed some components for better organization.
