+++ client/src/components/auth/AuthDashboard.jsx
