+++ client/src/components/flags/FlagInfoCard.jsx
@@ -8,7 +8,6 @@ import {
   DialogContentText,
   DialogTitle,
   Divider,
-  ListItem,
   Grid,
   Typography,
   Tooltip,
