+++ client/src/components/flags/FlagViewCircuitHealth.jsx
@@ -12,7 +12,7 @@ import apiClient from 'lib/apiClient';
 import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
-import CircuitCard from './CircuitCard';
+import CircuitCard from './FlagCircuitHealthCard';
 
 function FlagTimeSeries() {
   let { flagId } = useParams();
