--- client/src/components/home/Home.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Home() {
-  return <div>Home</div>;
-}
-
-export default Home;
