+++ client/src/components/logs/LogsDashboardHeader.jsx
